# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CircleTrade Agent Stack API
  description: |
    The Circle Agent Stack API exposes CircleTrade AI's export intelligence pipeline as
    individually callable micro-agent endpoints. Each endpoint maps to a distinct intelligence
    module: DDTRS (compliance & HS classification), MeridianFlow (route optimisation),
    and TradeMatch (market & tariff intelligence).

    **Authentication**: All `/api/*` endpoints require an `X-API-Key` header.
    Demo key: `ct-demo-key-2026`
  version: 1.0.0
  contact:
    name: CircleTrade AI
    url: https://circletrade.ai
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: product_description is required
        timestamp:
          type: string
          format: date-time
...
    TariffResponse:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        hs_code:
          type: string
        destination:
          type: string
        destination_code:
          type: string
        category:
          type: string
        tariff_rate:
          type: number
        currency:
          type: string
          example: percent
        tariff_label:
          type: string
          enum: [DUTY_FREE, LOW, MODERATE, HIGH]
        benchmark:
          type: object
          properties:
            lowest_in_category:
              type: object
              properties:
                destination: { type: string }
                rate: { type: number }
            highest_in_category:
              type: object
              properties:
                destination: { type: string }
                rate: { type: number }
        source:
          type: string
          example: TradeMatch
        timestamp:
          type: string
          format: date-time
    RouteResult:
      type: object
      properties:
        name: { type: string }
        score: { type: integer }
        transit_days: { type: integer }
        cost_index: { type: number }
        port_efficiency: { type: integer }
        political_risk: { type: string }
        regulatory_alignment: { type: string }
        primary_ports:
          type: array
          items: { type: string }
        recommendation: { type: string }
    RouteResponse:
      type: object
      properties:
        request_id: { type: string }
        query:
          type: object
          properties:
            origin: { type: string }
            destination: { type: string }
            product_category: { type: string, nullable: true }
        routes:
          type: array
          items:
            $ref: '#/components/schemas/RouteResult'
        best_route:
          $ref: '#/components/schemas/RouteResult'
        total_routes_found: { type: integer }
        source: { type: string }
        timestamp: { type: string, format: date-time }
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/tariff:
    post:
      summary: Look up import tariff rate
      tags: [TradeMatch]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hs_code, destination_code]
              properties:
                hs_code:
                  type: string
                  example: "6205.20"
                destination_code:
                  type: string
                  example: UAE
                  description: ISO 3-letter country code
                category:
                  type: string
                  example: textiles
                  description: Optional — inferred from hs_code if not provided
      responses:
        '200':
          description: Tariff lookup result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TariffResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Destination not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/route:
    post:
      summary: Optimise trade corridor / shipping route
      tags: [MeridianFlow]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [origin, destination]
              properties:
                origin:
                  type: string
                  example: UK
                destination:
                  type: string
                  example: UAE
                product_category:
                  type: string
                  example: textiles
      responses:
        '200':
          description: Route optimisation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...